# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farah API Documentation",
        "description": "توثيق واجهات برمجة التطبيقات (APIs) لمنصة فرح لتطبيق Flutter",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://farahbackend-production.up.railway.app",
            "description": "سيرفر الإنتاج (Railway Production)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "السيرفر المحلي (Local Server)"
        }
    ],
    "paths": {
...
        "/api/provider/verification": {
            "get": {
                "tags": [
                    "Provider"
                ],
                "summary": "عرض حالة توثيق حساب مزود الخدمة",
                "operationId": "1a755a15110bea08f138a864f2afa406",
                "responses": {
                    "200": {
                        "description": "حالة التوثيق"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/dashboard": {
            "get": {
                "tags": [
                    "Provider"
                ],
                "summary": "لوحة تحكم مزود الخدمة",
                "operationId": "7e20ae4c3274e2ca4839242300910409",
                "responses": {
                    "200": {
                        "description": "إحصاءات المزود والطلبات الأخيرة"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/onboarding": {
            "get": {
                "tags": [
                    "Provider"
                ],
                "summary": "عرض بيانات تسجيل مزود الخدمة",
                "operationId": "94b2c6a36c34e781434b1fc5d663aab1",
                "responses": {
                    "200": {
                        "description": "بيانات المزود"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Provider"
                ],
                "summary": "حفظ بيانات تسجيل مزود الخدمة",
                "operationId": "af9d3c6f4c16cccce8e37c9a19a5f0fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "business_name",
                                    "category_id",
                                    "description",
                                    "city_id",
                                    "address"
                                ],
                                "properties": {
                                    "business_name": {
                                        "type": "string",
                                        "example": "صالة أورنا"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "قاعة مناسبات مجهزة بالكامل"
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "غزة - الرمال"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "identity_document": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "commercial_register": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "cover_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم حفظ البيانات"
                    },
                    "422": {
                        "description": "بيانات غير صالحة"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/services": {
            "get": {
                "tags": [
                    "Provider"
                ],
                "summary": "عرض خدمات المزود",
                "operationId": "ecd745c240a22c827fc880a07ca7c67d",
                "responses": {
                    "200": {
                        "description": "قائمة الخدمات"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Provider"
                ],
                "summary": "إنشاء خدمة للمزود",
                "operationId": "6a0a0b60cc10a88ce703f93c64892ceb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "صالة أورنا"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "features": {
                                        "type": "string",
                                        "example": "[\"مواقف سيارات\",\"نظام صوتي\"]"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "area": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "booking_slots": {
                                        "type": "string",
                                        "example": "[{\"label\":\"الفترة المسائية\",\"start_time\":\"17:00\",\"end_time\":\"22:00\"}]"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "ILS"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم إنشاء الخدمة"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/services/{id}": {
            "post": {
                "tags": [
                    "Provider"
                ],
                "summary": "تعديل خدمة للمزود",
                "operationId": "b16ed5c514e42d99413c506a9b644481",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "features": {
                                        "type": "string"
                                    },
                                    "booking_slots": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تعديل الخدمة"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Provider"
                ],
                "summary": "حذف خدمة للمزود",
                "operationId": "5412933a7ae0c623c4d6fc6eb5a7b11c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم حذف الخدمة"
                    },
                    "403": {
                        "description": "المستخدم ليس مزود خدمة"
                    },
                    "404": {
                        "description": "الخدمة غير موجودة"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/services/{id}/visibility": {
            "patch": {
                "tags": [
                    "Provider"
                ],
                "summary": "إظهار أو إخفاء خدمة",
                "operationId": "67194fd54405a0cf1e77f29a1635cdd8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_available"
                                ],
                                "properties": {
                                    "is_available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تحديث ظهور الخدمة"
                    },
                    "404": {
                        "description": "الخدمة غير موجودة"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/services/{id}/availability": {
            "get": {
                "tags": [
                    "Provider"
                ],
                "summary": "عرض تقويم إتاحة الخدمة",
                "operationId": "5472492b02a7a54c78203102e1316942",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2026-09"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تقويم الخدمة"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Provider"
                ],
                "summary": "تحديث تقويم إتاحة الخدمة",
                "operationId": "17e874ccafa803544c47434ebdb1966e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "slots": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "unavailable_dates": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تحديث التقويم"
                    },
                    "422": {
                        "description": "بيانات غير صالحة"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/reviews": {
            "get": {
                "tags": [
                    "Provider"
                ],
                "summary": "عرض تقييمات خدمات المزود",
                "operationId": "3e2a25d52aa3805ba7ea6aad4636a2a9",
                "responses": {
                    "200": {
                        "description": "قائمة تقييمات العملاء"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "أدخل التوكن بهذا الشكل: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}